% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_dataset.R
\name{fetch_dataset}
\alias{fetch_dataset}
\title{Fetch dataset with dataset_id `dataset_id`.}
\usage{
fetch_dataset(dataset_id, restrictions = list(), vintage = "latest",
  dimensions = "code")
}
\arguments{
\item{dataset_id}{The dataset id you want to fetch. You can find dataset codes
on the Aeconomica website in the top right of the page showing the dataset.}

\item{restrictions}{(optional) A list with keys (corresponding to names of
dimensions) mapping to a vector of codes. Only series that have those codes
for that dimension are returned in the dataset.}

\item{vintage}{(default "latest") Specify a vintage for the series in the
dataset to fetch. Cannot be used if your specify vintages for individual series.
Vintage can be any of `current` (alias `latest`, default), `previous` or a
date in YYYY-MM-DD form.}

\item{dimensions}{can be one of "code" or "name." This dictates whether the
values for each dimension of the dataset are returned as the codes for each
dimension, or the long names that correspond to those codes.}
}
\description{
Fetch dataset with dataset_id `dataset_id`.
}
\examples{
\dontrun{
fetch_dataset("WJP_STATES")
}

}
