% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_series.R
\name{fetch_series}
\alias{fetch_series}
\title{Fetch a series, or list of series, of economic data from the Aeconomica API.}
\usage{
fetch_series(series, vintage = "latest")
}
\arguments{
\item{series}{A series, or number of series to fetch. Can be either a single
series code, a vector of series codes, or a list of series codes and vintage
pairs (with the series code as the name and the vintage as the value). You
can find series codes on the Aeconomica website in the top right of the page
showing the series.}

\item{vintage}{(optional) Specify a vintage for all of the series you want
to fetch. Cannot be used if your specify vintages for individual series.
Vintage can be any of `current` (alias `latest`, default), `previous` or a
date in YYYY-MM-DD form.}
}
\description{
Fetch a series, or list of series, of economic data from the Aeconomica API.
}
\examples{
\dontrun{
# Downloading a single series
fetch_series("WPI")

# Downloading a number of series
fetch_series(c("GDP_YE", "CPI_YE"))

# Use a list to download a different series with different vintages
fetch_series(list("GDP_YE" = "latest", "CPI_YE" = "previous"))
}

}
