% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_series_at_at.R
\name{fetch_series_as_at}
\alias{fetch_series_as_at}
\title{Fetch a series, or list of series, of economic data from the Aeconomica API.}
\usage{
fetch_series_as_at(series, as_at_date)
}
\arguments{
\item{series}{A series, or number of series to fetch. Can be either a single
series code, a vector of series codes, or a list of series codes and vintage
pairs (with the series code as the name and the vintage as the value). You
can find series codes on the Aeconomica website in the top right of the page
showing the series.}

\item{as_at_date}{Fetch the series vintage that was the latest vintage of
YYYY-MM-DD form.}
}
\description{
Fetch a series, or list of series, of economic data from the Aeconomica API.
}
\examples{
\dontrun{
# Downloading a single series as it was 23 September 2021
fetch_series_as_at("WPI", "2021-09-23")

# Downloading a number of series
fetch_series_as_at(c("GDP_YE", "CPI_YE"), "21-09-23")

# Use a list to download a different series with different `as_at_date`s
fetch_series_as_at(list("GDP_YE" = "21-09-23", "CPI_YE" = "2020-12-23"))
}

}
