% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_api.R
\name{set_apikey}
\alias{set_apikey}
\title{Set API key to access Aeconomica API}
\usage{
set_apikey(apikey)
}
\arguments{
\item{apikey}{The API key for your Aeconomica account.}
}
\description{
Visit your Aeconomica account page to find your API key.
}
